Move compiler types and exceptions to separate files, extract cairo1 logic to separate compiler
