+++ protostar/compiler/__init__.py
@@ -1,9 +1,11 @@
 from .compiled_contract_writer import CompiledContractWriter
 from .project_cairo_path_builder import ProjectCairoPathBuilder
 from .cairo0_project_compiler import (
-    CompilationException,
     LinkedLibrariesBuilder,
     Cairo0ProjectCompiler,
-    ProjectCompilerConfig,
-    SourceFileNotFoundException,
 )
+from .project_compiler_exceptions import (
+    SourceFileNotFoundException,
+    CompilationException,
+)
+from .project_compiler_types import ProjectCompilerConfig
